+++ changelog.md
@@ -1,5 +1,6 @@
 ## Changelog next version
 * Fixed typo in `CatchupSupportingBlockingSubscriptionConfig`, renamed method `dontSubscriptionPositionStorage` to `dontUseSubscriptionPositionStorage`.
+* Added `getSubscriptionPosition()` to `PositionAwareCloudEvent` that returns `Optional<SubscriptionPosition>`.
 
 
 ## Changelog 0.2.0 (2020-10-31)